Prepare simplification of repl definition
It seems convoluted that the users have to:
  a) create a repl definition to an arbitrary name;
  b) refer to that arbitrary name to get the definition;

This seems to cause a lot of confusion.
The new structure should allow for a direct reference:

{
  repl_definition = {
    python = {
      command = {"ipython3", "--whatever-flags"}
    }
  }
}

So the caller code will look up in the config and, if no
value is explicitly defined, it will lookup with the metamethod
to get the first available repl definition.

Explicitly defining the preferred repl definition
would be as easy as:

{
  repl_definition = {
    python = require("iron.fts.python").ipython
  }
}