Apply small fixes to internal functions
They should be self-contained and, as the unix philosophy mandates,
do one thing and do it well.

Therefore, we remove the indenting code (that apparently does nothing)
from `send_to_repl`.

Also, it seems logical to me that low level functions should not call
themselves. This lifts the work to the caller but makes them dead
simple.

As a final addition to this, users can configure to use scratch buffers
for the repl, which might make them lose the ability to toggle, but it
is configurable nonetheless.