Reeplace `if_repl_exists` function
It shouldn't do this much. Instead, it does a good job encapsulating the
condition checking, but it doesn't need to wrap the if check.

This is needed so we simplify the logic outside by reducing the number
of functions. Also, it reduces the number of calls to `ll.get` inside
this level. This is important because lowlevel shoud only take raw
arguments - it is low level after all - so the upper level should call
`ll.get` instead.