Take `meta` as argument
This further expans the concept of lowlevel as a series of functions
that do not interact directly, but the wiring is done by the upper
level.

This is important so we can always send data to any repl. While the
default usage will ensure we always send to the appropriate one,
this is not concern of the low level and exceptions can be made - i.e.
closing all repls.