Refactor ll memory management (#227)
* Reeplace `if_repl_exists` function

It shouldn't do this much. Instead, it does a good job encapsulating the
condition checking, but it doesn't need to wrap the if check.

This is needed so we simplify the logic outside by reducing the number
of functions. Also, it reduces the number of calls to `ll.get` inside
this level. This is important because lowlevel shoud only take raw
arguments - it is low level after all - so the upper level should call
`ll.get` instead.

* Take `meta` as argument

This further expans the concept of lowlevel as a series of functions
that do not interact directly, but the wiring is done by the upper
level.

This is important so we can always send data to any repl. While the
default usage will ensure we always send to the appropriate one,
this is not concern of the low level and exceptions can be made - i.e.
closing all repls.

* Remove some of the deprecated functions

* Replace usages of `if_repls_exists`

* Raise `error` when ft is nil

This allows for better flow control and should simplify callers logic

* Cleanup `ft == nil` checks

Also, remove deprecated function

* Add ft to meta

So we can reuse it later

* Get filetype from meta

This removes all usages of `ll.get` from inside this layer

* Add importante note

* Ensure all scopes follow the same memory structure

So we can iterate through

* Remove variable declaration

So we don't shadow it

* Fix wrong var name